requirements.txt
absl-py==2.3.1
accelerate==1.11.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
archspec @ file:///home/conda/feedstock_root/build_artifacts/archspec_1737352602016/work
attrs==25.4.0
boltons @ file:///home/conda/feedstock_root/build_artifacts/boltons_1749686179973/work
Brotli @ file:///Users/runner/miniforge3/conda-bld/brotli-split_1749229914644/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1752493286596/work/certifi
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1725560578465/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1746214863626/work
click==8.3.0
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
conda @ file:///Users/runner/miniforge3/conda-bld/conda_1744220052803/work/conda-src
conda-libmamba-solver @ file:///home/conda/feedstock_root/build_artifacts/conda-libmamba-solver_1742219570693/work/src
conda-package-handling @ file:///home/conda/feedstock_root/build_artifacts/conda-package-handling_1736345463896/work
conda_package_streaming @ file:///home/conda/feedstock_root/build_artifacts/conda-package-streaming_1751548120229/work
datasets==4.3.0
dill==0.4.0
distro @ file:///home/conda/feedstock_root/build_artifacts/distro_1734729835256/work
evaluate==0.4.6
filelock==3.20.0
frozendict @ file:///Users/runner/miniforge3/conda-bld/frozendict_1728841362247/work
frozenlist==1.8.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
grpcio==1.76.0
h11==0.16.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1738578511449/work
hf-xet==1.2.0
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
Jinja2==3.1.6
joblib==1.5.2
jsonpatch @ file:///home/conda/feedstock_root/build_artifacts/jsonpatch_1733814567314/work
jsonpointer @ file:///Users/runner/miniforge3/conda-bld/jsonpointer_1725302931775/work
libmambapy @ file:///Users/runner/miniforge3/conda-bld/mamba-split_1746516032751/work/libmambapy
Markdown==3.9
MarkupSafe==3.0.3
menuinst @ file:///Users/runner/miniforge3/conda-bld/menuinst_1753546295087/work
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.16
networkx==3.5
nltk==3.8.1
numpy==2.3.4
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas==2.3.3
pillow==12.0.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1746710438/work
pluggy @ file:///home/conda/feedstock_root/build_artifacts/pluggy_1747339660894/work
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.2
pyarrow==22.0.0
pycocoevalcap==1.2
pycocotools==2.0.10
pycosat @ file:///Users/runner/miniforge3/conda-bld/pycosat_1732588505673/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pydantic==2.12.3
pydantic_core==2.41.4
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.10.23
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1749498106507/work
rouge_score==0.1.2
ruamel.yaml @ file:///Users/runner/miniforge3/conda-bld/ruamel.yaml_1749480012856/work
ruamel.yaml.clib @ file:///Users/runner/miniforge3/conda-bld/ruamel.yaml.clib_1728724504988/work
safetensors==0.6.2
sentry-sdk==2.42.1
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tokenizers==0.22.1
torch==2.9.0
torchvision==0.24.0
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1735661334605/work
transformers==4.57.1
truststore @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_truststore_1739009763/work
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1750271362675/work
wandb==0.22.2
Werkzeug==3.1.3
wheel==0.45.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.23.0
